Make issues found exit non-zero (2)
Why 2? Because 1 is used for help string. But this way one can use the
tool in a test script (for instance) and have it exit non zero for any
issues found.

I did not make it an unsigned long long but rather an unsigned long just
for a bit more portability (even though unsigned long long should be
fine in 2023) and because it should not matter anyway. I incremented the
counter but having it be a short int and just setting it to 1 would also
suffice.